pyproject.toml
[project]
name = "se-imr-mtl-kd"
version = "0.1.0"
description = "Speech Emotion & Intensity Multi-Task Learning with Knowledge Distillation"
authors = [
    {name = "Liu Yang", email = "yang.liu6@siat.ac.cn"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fairseq>=0.12.2",
    "huggingface-hub>=0.34.4",
    "hyperpyyaml>=1.2.2",
    "librosa>=0.11.0",
    "matplotlib>=3.10.6",
    "nvitop>=1.5.3",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "pysndfx>=0.3.6",
    "pyyaml>=6.0.2",
    "speechbrain>=1.0.3",
    "torch>=2.3.1",
    "torchaudio>=2.3.1",
    "torchvision>=0.18.1",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
]

[project.urls]
Homepage = "https://github.com/Hex4C59/SE_IMR_MTL_KD"
Repository = "https://github.com/Hex4C59/SE_IMR_MTL_KD.git"





